test: update stale test imports for renamed *-service.ts files
Three tests (cli-skills, cli-sidecar-path, slash-commands) imported
the pre-rename paths services/{cli,cli-skills}.ts. Repointed at the
new *-service.ts filenames so the full vitest suite passes again.
